refactor: move Validation to a protocols folder
