+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:45177/7664bdbf-b85f-4619-ac33-54c852019ce5?"}
+{"mongoUri":"mongodb://127.0.0.1:41521/be055c8e-0c29-4cc9-a18c-09a705185686?"}
