+++ src/main/factories/signup/signup-validation.ts
@@ -1,5 +1,5 @@
 import { RequiredFielValidation } from '../../../presentation/helpers/validators/required-field-validation'
-import { Validation } from '../../../presentation/helpers/validators/validation'
+import { Validation } from '../../../presentation/protocols/validation'
 import { ValidationComposite } from '../../../presentation/helpers/validators/validation-composite'
 import { CompareFieldsValidation } from '../../../presentation/helpers/validators/compare-field-validation'
 import { EmailValidatorAdapter } from '../../../utils/email-validator-adapter'
