+++ src/presentation/controllers/login/login-protocols.ts
@@ -1,4 +1,4 @@
 export * from '../../protocols'
 export * from '../../protocols/email-validator'
 export * from '../../../domain/useCases/authentication'
-export * from '../../helpers/validators/validation'
+export * from '../../protocols/validation'
