+++ src/presentation/controllers/signup/signup-protocols.ts
@@ -2,4 +2,4 @@ export * from '../../protocols'
 export * from '../../protocols/email-validator'
 export * from '../../../domain/useCases/add-account'
 export * from '../../../domain/models/account'
-export * from '../../helpers/validators/validation'
+export * from '../../protocols/validation'
